NABI-229--refactor : dib 리팩터링
