--- src/main/java/org/prgrms/nabimarketbe/domain/dibs/dto/response/DibResponseDTO.java
@@ -1,6 +1,0 @@
-package org.prgrms.nabimarketbe.domain.dibs.dto.response;
-
-public record DibResponseDTO<T>(
-	T dibInfo
-) {
-}
