+++ src/main/java/org/prgrms/nabimarketbe/domain/dibs/dto/response/wrapper/DibResponseDTO.java
@@ -1,0 +1,6 @@
+package org.prgrms.nabimarketbe.domain.dibs.dto.response.wrapper;
+
+public record DibResponseDTO<T>(
+	T dibInfo
+) {
+}
